+++ src/main/java/gift/product/controller/ProductViewController.java
@@ -1,13 +1,9 @@
 package gift.product.controller;
 
 import gift.category.service.CategoryService;
-import gift.option.domain.OptionDTO;
-import gift.option.service.OptionService;
-import gift.product.domain.Product;
+import gift.product.option.service.OptionService;
 import gift.product.domain.ProductDTO;
 import gift.product.service.ProductService;
-import java.util.ArrayList;
-import java.util.List;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
